[JSC] operationIterateResults needs to truncate i32 results https://bugs.webkit.org/show_bug.cgi?id=301211 rdar://153280651
Reviewed by Daniel Liu.

BBQ expects that values are truncated to their proper size when
producing the value. When converting from JS signed 32-bits to
wasm i32, should be zero extending rather than sign extending.

Test: JSTests/wasm/stress/i32-multivalue-truncate.js
* JSTests/wasm/stress/i32-multivalue-truncate.js: Added.
(async test.getTwoI32s):
(async test):
* Source/JavaScriptCore/wasm/WasmOperations.cpp:
(JSC::Wasm::JSC_DEFINE_NOEXCEPT_JIT_OPERATION):

Originally-landed-as: 301765.80@safari-7623-branch (dad39caa1d1c). rdar://166339272
Canonical link: https://commits.webkit.org/304409@main

+++ Source/JavaScriptCore/wasm/WasmOperations.cpp
@@ -1413,7 +1413,7 @@ JSC_DEFINE_JIT_OPERATION(operationIterateResults, void, (JSWebAssemblyInstance* 
         const auto& returnType = signature->returnType(index);
         switch (returnType.kind) {
         case TypeKind::I32:
-            unboxedValue = value.toInt32(globalObject);
+            unboxedValue = static_cast<uint32_t>(value.toInt32(globalObject));
             break;
         case TypeKind::I64:
             unboxedValue = value.toBigInt64(globalObject);
